Move LAContextProtocol to separate file
